+++ android/src/main/java/com/nefta/sdk/IListener.java
@@ -1,0 +1,5 @@
+package com.nefta.sdk;
+
+public interface IListener {
+    void Invoke(int responseId, int adapterResponseType, String adapterResponse);
+}